feat : GruzMother State Machine
GruzMother State Machine Die제외 구현

+++ Assets/Resources/Data/PlayerData.cs
@@ -17,9 +17,9 @@ public class PlayerData : ScriptableObject
         public string description;
 
         public int mapHp;
-        public float curHp;
+        public int curHp;
         public int maxSoul;
-        public float curSoul;
+        public int curSoul;
         public int attackDamage;
         public int howlingDamage;
         public int shotSoulDamage;
